♻ Refactor ApiHelper
