+++ modules/apihelper/logger.py
@@ -1,0 +1,3 @@
+import logging
+
+logger = logging.getLogger("ErrorPush")
