+++ modules/apihelper/typedefs.py
@@ -1,4 +1,6 @@
 from typing import Dict, Any
 
+__all__ = ("POST_DATA", "JSON_DATA")
+
 POST_DATA = Dict[str, Any]
 JSON_DATA = Dict[str, Any]
